+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 This file documents all notable changes to Aerospike Prometheus Exporter
 
 
+## [v1.5.0](https://github.com/aerospike/aerospike-prometheus-exporter/releases/tag/v1.5.0)
+
+### Features
+- [TOOLS-1946] - Add support for per-job scan and query statistics
+- [TOOLS-1947] - Add support for secondary index statistics
+
+
 ## [v1.4.1](https://github.com/aerospike/aerospike-prometheus-exporter/releases/tag/v1.4.1)
 
 ### Improvements